some

+++ main-service/src/main/java/ru/practicum/mainservice/category/service/CategoryServiceImpl.java
@@ -3,16 +3,21 @@ package ru.practicum.mainservice.category.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.mainservice.category.dto.CategoryDto;
 import ru.practicum.mainservice.category.dto.NewCategoryDto;
 import ru.practicum.mainservice.category.mapper.CategoryMapper;
 import ru.practicum.mainservice.category.model.Category;
 import ru.practicum.mainservice.category.repository.CategoryRepository;
+import ru.practicum.mainservice.event.model.Event;
+import ru.practicum.mainservice.event.repository.EventRepository;
 import ru.practicum.mainservice.exception.IncorrectFieldException;
 import ru.practicum.mainservice.exception.IncorrectObjectException;
-import ru.practicum.mainservice.util.CategoryChecker;
+import ru.practicum.mainservice.util.State;
+import ru.practicum.mainservice.util.checker.CategoryChecker;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -21,17 +26,18 @@ import java.util.stream.Collectors;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
     private final CategoryChecker categoryChecker;
     private final CategoryMapper mapper;
 
     @Override
-    public CategoryDto addCategory(NewCategoryDto newCategoryDto) {
+    public CategoryDto addCategory(NewCategoryDto newCategoryDto) throws IncorrectFieldException {
         Category newCategory = mapper.toCategory(newCategoryDto);
         try {
             newCategory = categoryRepository.save(newCategory);
             return mapper.toDto(newCategory);
         } catch (DataIntegrityViolationException ex) {
-            throw new DataIntegrityViolationException("Data of category exception");
+            throw new IncorrectFieldException("Data of category exception");
         }
     }
 
@@ -49,22 +55,32 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public CategoryDto updateCategory(CategoryDto categoryDto) throws IncorrectObjectException, IncorrectFieldException {
-        categoryChecker.categoryExist(categoryDto.getId());
-        categoryChecker.idIsNotBlank(categoryDto.getId());
-        final Category updatedCategory = categoryRepository.getReferenceById(categoryDto.getId());
-        updatedCategory.setName(categoryDto.getName());
+    public CategoryDto updateCategory(Long catId, CategoryDto categoryDto) throws IncorrectObjectException, IncorrectFieldException {
+        categoryChecker.categoryExist(catId);
+        categoryChecker.idIsNotBlank(catId);
+        final Category updatedCategory = categoryRepository.getReferenceById(catId);
+        if (categoryDto.getName() != null) {
+            updatedCategory.setName(categoryDto.getName());
+        }
         try {
             return mapper.toDto(categoryRepository.save(updatedCategory));
         } catch (DataIntegrityViolationException e) {
-            throw new DataIntegrityViolationException("Name of category exception");
+            throw new IncorrectFieldException("Name of category exception");
         }
     }
 
     //TODO***
     @Override
-    public void deleteCategory(Long categoryId) throws IncorrectObjectException {
-        categoryChecker.categoryExist(categoryId);
+    public void deleteCategory(Long categoryId) throws IncorrectObjectException, IncorrectFieldException {
+        List<Event> events = eventRepository.findAllByCategoryIdInAndEventDateIsAfter(List.of(categoryId),
+                LocalDateTime.now(), Pageable.unpaged());
+        if (!events.isEmpty()) {
+            boolean isConfirmedState = events.stream()
+                    .anyMatch(e -> e.getState().equals(State.PUBLISHED));
+            if (!isConfirmedState) {
+                throw new IncorrectFieldException("Not delete. Category contains events");
+            }
+        }
         categoryRepository.deleteById(categoryId);
     }
 }
